requirements.txt
openai
einops
gradio
imageio
decord
simplet5
imageio-ffmpeg
Pillow
scipy
numpy
openmim
mmcv
spacy
simplet5
decord
omegaconf
lvis
scipy
boto3
jsonschema
entrypoints
nltk
webdataset
accelerate
bitsandbytes
langchain==0.0.101
timm==0.4.12
transformers
fairscale==0.4.4
pycocoevalcap
torch==1.13.1
torchvision==0.14.1
wget==3.2
setuptools==59.5.0